+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## X.X.X
+* Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'
+* Added a call to enroll users to A/B tests when getting all remote config values: 'getAllValuesAndEnroll'
+
 ## 23.8.2
 * Fixed APM bug where automatic foreground, background tracking would track wrong if the SDK was not initialized while the app was not in the foreground